pyproject.toml
[project]
name = "policyengine-household-api"
version = "0.13.14"
description = "PolicyEngine Household API"
readme = "README.md"
authors = [
    { name = "PolicyEngine", email = "hello@policyengine.org" },
]
requires-python = ">=3.12"
dependencies = [
    "anthropic",
    "Authlib>=1.3.1,<1.7.0",
    "cloud-sql-python-connector",
    "flask>=2.2",
    "flask-cors>=3",
    "flask-sqlalchemy>=3",
    "google-cloud-logging",
    "google-cloud-storage",
    "gunicorn",
    "inflect",
    "policyengine_canada==0.96.3",
    "policyengine-ng==0.5.1",
    "policyengine-il==0.1.0",
    "policyengine_uk==2.31.0",
    "policyengine_us==1.626.1",
    "policyengine_core<=3.23.6",
    "pyjwt",
    "Flask-Caching==2.0.2",
    "urllib3<=1.26.20,>=1.21.1",
    "python-dotenv",
    "pymysql",
    "Flask-Limiter",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest-timeout",
    "ruff>=0.9.0",
    "towncrier>=24.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
required-version = "==0.8.5"

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.towncrier]
package = "policyengine_household_api"
directory = "changelog.d"
filename = "CHANGELOG.md"
title_format = "## [{version}] - {project_date}"
issue_format = ""
underlines = ["", "", ""]

[[tool.towncrier.type]]
directory = "breaking"
name = "Breaking changes"
showcontent = true

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[tool.hatch.build.targets.wheel]
packages = ["policyengine_household_api"]
